requirements.txt
discord.py
google-genai
pydantic
python-dotenv
unidecode
aiosqlite
ruamel.yaml
